Moved adding donation to milestones to the state
