+++ src/main/java/asu/eng/gofund/controller/DonationController.java
@@ -39,8 +39,7 @@ public class DonationController {
     @Autowired
     private UserRepo userRepo;
 
-    @Autowired
-    private MilestonesRepo milestonesRepo;
+
 
     DonationView donationView = new DonationView();
     CoreView coreView = new CoreView();
@@ -104,22 +103,7 @@ public class DonationController {
         return campaignView.redirectToCampaignWithID(campaignId);
     }
 
-    private void addDonationToMileStones(Long campaignId, Donation decoratedDonation) {
-        List<Milestone> milestones = milestonesRepo.getAllByCampaignId(campaignId);
-        double tempAmount = decoratedDonation.getAmount();
-        for (Milestone milestone : milestones) {
-            if (tempAmount <= 0) break;
-            double remainingAmount = milestone.getTargetAmount() - milestone.getCurrentFunds();
-            if (tempAmount >= remainingAmount) {
-                milestone.setCurrentFunds(milestone.getTargetAmount());
-                tempAmount -= remainingAmount;
-            } else {
-                milestone.setCurrentFunds(milestone.getCurrentFunds() + tempAmount);
-                tempAmount = 0;
-            }
-            milestonesRepo.save(milestone);
-        }
-    }
+
 
     private Donation handleDifferentCurrency(Donation donation, Campaign campaign, boolean refunding) {
         if (campaign.getCurrency() == CustomCurrency.USD && donation.getCurrency() == CustomCurrency.EGP) {
